% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{ci_cor}
\alias{ci_cor}
\title{Confidence interval of sample correlation coefficient}
\usage{
ci_cor(alpha, rho, n)
}
\arguments{
\item{alpha}{confidence level}

\item{rho}{population correlation coefficient}

\item{n}{sample size}
}
\description{
Based on the closed form solution in Provost (2015). May fail for large n (>200).
Not suitable for constructing confidence intervals for subset correlation coefficients, i.e. correlation
coefficients from finite populations.
}

