% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{ci_cor_fisher}
\alias{ci_cor_fisher}
\title{Confidence interval of sample correlation coefficient (also in finite populations)}
\usage{
ci_cor_fisher(alpha, rho, n, Npop)
}
\arguments{
\item{alpha}{confidence level}

\item{rho}{population correlation coefficient}

\item{n}{sample size}

\item{Npop}{Population size}
}
\description{
Based on Fisher's z-transformation and a correction of the standard deviation for
finite samples: (Npop - n) / (Npop - 1)
}

