% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{dcorr}
\alias{dcorr}
\title{Density function of sample correlation coefficient based on the closed form
solution in Provost (2015)}
\usage{
dcorr(r, n, rho)
}
\arguments{
\item{r}{Value of the correlation coefficient}

\item{n}{Sample size}

\item{rho}{True correlation in the population}
}
\description{
Uses the hypergeometric function 2F1 from the gsl package.
}

