% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{E_r_integral}
\alias{E_r_integral}
\title{Expected value of sample correlation coefficient}
\usage{
E_r_integral(n, rho)
}
\arguments{
\item{n}{Sample size}

\item{rho}{True correlation in the population}
}
\description{
Calculated as the integral of r times the distribution function dcorr in [-1, 1].
}

