% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{E_z}
\alias{E_z}
\title{Functions for Fisher's z-transformation}
\usage{
E_z(rho)
}
\arguments{
\item{rho}{Population correlation coefficient}

\item{r}{Sample correlation coefficient}

\item{n}{Sample size}

\item{Npop}{Population size in case of finite populations, e.g. when a subset was
extracted from a sample}
}
\description{
Convert a sample correlation coefficient (r) to Fisher's z or the other way
around, compute the expected value of z (E_z), compute the standard error of z
(se_z), and the standard error of z with a correction for finite samples (se_z_finite) with
Npop being the size of the full sample.
}
\seealso{
\code{\link{rtoz}}, \code{\link{ztor}}, \code{\link{E_z}}, \code{\link{se_z}}, \code{\link{se_z_finite}}
}

