% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getBiCop}
\alias{getBiCop}
\title{Simulate bivariate normally distributed data with specified correlation, mean and sd}
\usage{
getBiCop(N, rho, rho_sub = NA, fraction = NA, mar.fun = rnorm, x = NULL,
  ...)
}
\arguments{
\item{rho}{True correlation coefficient in the population. If rho_sub is given,
this is the correlation in the rest of the sample without the subset.}

\item{rho_sub}{True correlation coefficient in the subset. Default is no
seperate subset.}

\item{fraction}{The size of the subset as a percentage of the full sample.
The subset is then contained in the first round(N * fraction) rows of the
returned data frame.}

\item{n}{Sample size.}
}
\description{
The variables x and y stem from a distribution with correlation rho, but the
actual correlation between x and y will vary around that true value.
}
\details{
Code from http://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variable
}

