% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getBiCop_indep}
\alias{getBiCop_indep}
\title{Simulate independent normally distributed data}
\usage{
getBiCop_indep(N, mu1 = 0, mu2 = 0, sd1 = 1, sd2 = 1)
}
\arguments{
\item{mu1}{Mean of x}

\item{mu2}{Mean of y}

\item{sd1}{Sd of x}

\item{sd2}{Sd of y}

\item{n}{Sample size}
}
\description{
Simulate independent normally distributed data
}

