% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getBiCop_nonnor}
\alias{getBiCop_nonnor}
\title{Simulate  non-normally distributed data with specified correlation, mean and sd}
\usage{
getBiCop_nonnor(N, rho, rho_sub = NA, fraction = NA, mu1 = 0, mu2 = 0,
  sd1 = 1, sd2 = 1, skew1 = -1, skew2 = 1, kurtosis1 = 1,
  kurtosis2 = 1, rhotol = 0.1)
}
\arguments{
\item{N}{Size of generated data (rows)}

\item{rho}{Target correlation coefficient}

\item{rho_sub}{True correlation coefficient in the subset. Default is no
seperate subset.}

\item{fraction}{The size of the subset as a percentage of the full sample.
The subset is then contained in the first round(N * fraction) rows of the
returned data frame.}

\item{skew1}{Skew of first variable}

\item{skew2}{Skew of second variable}

\item{kurtosis1}{Kurtosis of first variable}

\item{kurtosis2}{Kurtosis of second variable}

\item{m1}{Mean of first variable}

\item{m2}{Mean of second variable}

\item{s1}{Standard deviation of first variable}

\item{s2}{Standard deviation of second variable}
}
\description{
Draws random numbers from distributions with specified skew and kurtosis based on
PoisNonNor::RNG_P_NN. Checks the obtained correlation and issues a warning if it
deviates more than rhotol from rho.
}

