% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{inference_cor_perm}
\alias{inference_cor_perm}
\title{Calculate confidence intervals and p-value for a given subsample correlation}
\usage{
inference_cor_perm(x, y, type = "simple", n_permut = 1000,
  inSubset = NULL)
}
\arguments{
\item{x}{x variable (numeric vector)}

\item{y}{y variable (numeric vector)}

\item{type}{"simple" or "studentized" permutation test}

\item{n_permut}{number of permutations for the permutation test}

\item{inSubset}{integer vector of the row numbers of all observations that belong
to the subset}
}
\description{
Given a data set (x and y) and a vector that contains the row numbers of the
observations that belong to the subset (inSubset) the functions runs n_permut permutations
for determining 95% and 99% confidence intervals and the significance of the
given subset correlation coefficient (r)
}

