% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{makeSteps}
\alias{makeSteps}
\title{Create sequence of integers with increasing step size}
\usage{
makeSteps(nmax, stepPerc = 0.01, initialSamplesize = 10, manValues = NULL)
}
\arguments{
\item{nmax}{largest and last value of sequence}

\item{stepPerc}{every element of the sequence is at least stepPerc percent larger
than the preceding one}

\item{initialSamplesize}{start of the sequence}

\item{manValues}{manually given values that are added to the sequence even if they
do not fit into the step size that is defined by stepPerc}
}
\description{
The sequence starts at initialSamplesize and every next element
is (at least, due to rounding) increased by stepPerc percent. The last
element is always nmax.
}

