% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{moments_R_even}
\alias{moments_R_even}
\title{Moment functions for the sample correlation coefficient based on the solutions
of Anderson (via Provost, 2015)}
\usage{
moments_R_even(rho, n, k, maxterms = 1000)
}
\arguments{
\item{rho}{True correlation coefficient in the population}

\item{n}{Sample size}

\item{k}{Moment number}
}
\description{
Stops the calculation of the sum term as soon as an infinite value is
returned, which may happen very soon or immediately for large n. In that
case the results are not reliable.
}

