% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{pcorr}
\alias{pcorr}
\title{Distribution function of sample correlation coefficient based on the closed form
solution in Provost (2015)}
\usage{
pcorr(q, n, rho, lower.tail = TRUE)
}
\arguments{
\item{q}{Value of the correlation coefficient}

\item{n}{Sample size}

\item{rho}{True correlation in the population}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}
}
\description{
Uses the hypergeometric function 2F1 from the gsl package.
Calculates the distribution directly as the integral of dcorr().
}

