% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{permute_ecdf}
\alias{permute_ecdf}
\title{Combined function for p-value and CI of a sample correlation coefficient
based on Fisher's z-transformation}
\usage{
permute_ecdf(x, y, fraction = NULL, n_subset = NULL, n_permut = 1000)
}
\arguments{
\item{x}{Variable x, same length as y}

\item{y}{Variable y, same length as x}

\item{fraction}{Size of the subset in relation to the full set in percent}

\item{n_subset}{Size of the subset}

\item{n_permut}{number of permutations}

\item{alpha}{confidence level for confidence interval}

\item{rho}{population correlation coefficient}

\item{n}{sample size}

\item{Npop}{Population size}

\item{alternative}{For p-value: a character string specifying the alternative hypothesis,
 must be one of "two.sided" (default), "greater" or "less"
draw a permutation from the full sample with size n_subset
save empirical distribution of cor_subset

Confidence intervals can be derived using quantile()
p-values of observed cor_subset can be derived using ecdf()
Either fraction or n_subset should be specified.}
}
\description{
Based on Fisher's z-transformation and a correction of the standard deviation for
finite samples: (Npop - n) / (Npop - 1)
Calculates probability of observing a sample correlation coefficient r given
a sample size n, a population size Npop and the population correlation coefficient rho
as the p-value and also outputs the confidence interval from ci_cor_fisher().
}

