% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{pval_cor}
\alias{pval_cor}
\title{p-value of sample correlation coefficient}
\usage{
pval_cor(r, rho, n, alternative = "greater")
}
\arguments{
\item{r}{sample correlation coefficient}

\item{rho}{population correlation coefficient}

\item{n}{sample size}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "greater" (default) or "less"}
}
\description{
Based on the closed form solution in Provost (2015). May fail for large n (>200).
Calculates probability of observing a sample correlation coefficient r given
a sample size n and the population correlation coefficient rho. Not suitable
for assessing the probability of subset correlation coefficients, i.e. correlation
coefficients from finite populations.
}

