% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{rtoz}
\alias{rtoz}
\title{Functions for Fisher's z-transformation}
\usage{
rtoz(r)
}
\arguments{
\item{r}{Sample correlation coefficient}

\item{rho}{Population correlation coefficient}

\item{n}{Sample size}

\item{Npop}{Population size in case of finite populations, e.g. when a subset was
extracted from a sample}
}
\description{
Functions for Fisher's z-transformation
}

