% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{run_cor_sim}
\alias{run_cor_sim}
\title{Simulate repetitions of drawing a subset from data of a specified type,
 calculate the (subset) correlation and compute coverage probabilities and
 type 1 errors for specified inference type}
\usage{
run_cor_sim(N, fraction = NA, rho = 0, rho_sub = NA,
  distribution = "norm_norm", test = "z", permuttype = "simple",
  n_permut = 1000, repetitions = 100, progress_bar = T)
}
\arguments{
\item{N}{(integer) Size of the full sample}

\item{fraction}{Size of the subset as a percentage of the full set}

\item{distribution}{One of "norm_norm", "log_norm" or "log_log". Refers to
the "type" argument in getBiCop_exact().}

\item{test}{"z" for Fisher's z transformation and p-values and confidence
intervals based on normal theory. "permutation" for running a permutation
test}

\item{permuttype}{"simple" or "studentized"}

\item{n_permut}{Number of permutations for the permutation tests}

\item{repetitions}{The simulation will be repeated this many times}
}
\description{
Simulate repetitions of drawing a subset from data of a specified type,
 calculate the (subset) correlation and compute coverage probabilities and
 type 1 errors for specified inference type
}

