% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{run_sim_fisher}
\alias{run_sim_fisher}
\title{Simulate repetitions of drawing a subset from data of a specified type,
 calculate the (subset) correlation and compute coverage probabilities and
 type 1 errors for specified inference type}
\usage{
run_sim_fisher(N, fraction, rho = 0, rho_sub = NA,
  distribution = "normal", kurtosis1 = 1, kurtosis2 = 1, skew1 = 1,
  skew2 = 1, repetitions = 100, progress_bar = T, summarize = F)
}
\arguments{
\item{N}{(integer) Size of the full sample}

\item{fraction}{Size of the subset as a percentage of the full set}

\item{distribution}{One of "normal" or "nonnormal". The type of distribution the
variables are simulated from. In case of "nonnormal" the skewness and kurtosis
can deviate from a normal distribution}

\item{kurtosis1}{Kurtosis of the distribution of x}

\item{kurtosis2}{Kurtosis of the distribution of y}

\item{skew1}{Skewness of the distribution of x}

\item{skew2}{Skewness of the distribution of y}

\item{repetitions}{The simulation will be repeated this many times}
}
\description{
Simulate repetitions of drawing a subset from data of a specified type,
 calculate the (subset) correlation and compute coverage probabilities and
 type 1 errors for specified inference type
}

