% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{sim_subcor}
\alias{sim_subcor}
\title{Simulate a single subset correlation of a randomly drawn subset of specified size}
\usage{
sim_subcor(dat, fraction, sampling = T)
}
\arguments{
\item{dat}{data.frame or matrix with two columns which include the data}

\item{fraction}{Size of the subset as percentage of nrow(dat), e.g. 0.3}

\item{sampling}{(logical) If TRUE, sample round(nrow(dat) * fraction) rows from
the data to calculate the subset correlation. Otherwise take the first
round(nrow(dat) * fraction) rows}
}
\description{
Simulate a single subset correlation of a randomly drawn subset of specified size
}

