% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{var_r_integral}
\alias{var_r_integral}
\title{Variance of sample correlation coefficient}
\usage{
var_r_integral(n, rho)
}
\arguments{
\item{n}{Sample size}

\item{rho}{True correlation in the population}
}
\description{
Calculated as the integral of (r - E(r)) times the distribution function dcorr in [-1, 1].
Thus based on the closed form solution in Provost (2015).
}

