% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{getBiCop_exact}
\alias{getBiCop_exact}
\title{Simulate bivariate normally distributed data with specified correlation, mean and sd}
\usage{
getBiCop_exact(N, rho, rho_sub = NA, mu1 = 0, sd1 = 1, fraction = NA)
}
\arguments{
\item{N}{Sample size}

\item{rho}{True correlation coefficient in the sample}

\item{mu1}{Expected value of the mean of variable x}

\item{sd1}{Expected value of the standard deviation of variable x}
}
\description{
The correlation between x and y will be exactly rho.
}
\details{
Code from http://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variable
}

