% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{permute_ecdf}
\alias{permute_ecdf}
\title{Draw a permutation from the full sample with size n_subset and save empirical distribution of cor_subset}
\usage{
permute_ecdf(x, y, fraction = NULL, n_subset = NULL, n_permut = 1000)
}
\arguments{
\item{x}{Variable x, same length as y}

\item{y}{Variable y, same length as x}

\item{fraction}{Size of the subset in relation to the full set in percent}

\item{n_subset}{Size of the subset}

\item{n_permut}{number of permutations}
}
\description{
Confidence intervals can be derived using quantile()
p-values of observed cor_subset can be derived using ecdf()
Either fraction or n_subset should be specified.
}

